+++ src/features/calendar.rs
@@ -368,6 +368,7 @@ impl DateFormat {
             raw_having_conditions: Vec::new(),
             raw_join_conditions: Vec::new(),
             raw_aggregations: Vec::new(),
+            uses_group_by_all: false
         })
     }
 
@@ -587,5 +588,6 @@ impl DateFormat {
             raw_having_conditions: Vec::new(),
             raw_join_conditions: Vec::new(),
             raw_aggregations: Vec::new(),
+            uses_group_by_all: false
         })
     }